+++ core/jvm/src/main/scala/zio/sql/Sql.scala
@@ -1,8 +1,9 @@
 package zio.sql
 
-import zio.sql.rendering.Rendering
+import zio.sql.rendering.{ RenderModule, Rendering }
 
-trait Sql extends SelectModule with DeleteModule with UpdateModule with ExprModule with TableModule { self =>
+trait Sql extends SelectModule with DeleteModule with UpdateModule with ExprModule with TableModule with RenderModule {
+  self =>
 
   /*
    * (SELECT *, "foo", table.a + table.b AS sum... FROM table WHERE cond) UNION (SELECT ... FROM table)
@@ -17,7 +18,7 @@ trait Sql extends SelectModule with DeleteModule with UpdateModule with ExprModu
    * SELECT ARBITRARY(age), COUNT(*) FROM person GROUP BY age
    */
 
-  type SqlRendering[A] <: Rendering[A]
+  type SqlRenderer[-A] <: Rendering[A]
 
   def select[F, A, B <: SelectionSet[A]](selection: Selection[F, A, B]): SelectBuilder[F, A, B] =
     SelectBuilder(selection)
